New introduction section, providing better context for the purpose of the testbed and its experiments and clearer separation between the Internet's Root Server System and the Yeti DNS Project.
Replaced the "Problem Statement" with an "Areas of Study" section.
Re-structured each broad category in the form of a list of questions
rather than prose. Removed the more explicit commentary on the
number of root servers since that is expected to be seen in a
political light, rather than the technical tone intended.